REFACTOR : 쿼리 폴더 구조 수정
