+++ src/hooks/queries/product-detail/useDeleteMutation.ts
