+++ src/hooks/queries/product-detail/useHideMutation.ts
