+++ src/hooks/queries/product-detail/useOnSaleMutation.ts
